requirements/base.txt
Django==1.7.3
django-cors-headers==1.0.0
dj-database-url==0.3.0
django-timezone-field==1.1
djangorestframework==3.0.3
nameparser==0.3.3
Pillow==2.7.0
psycopg2==2.5.4
python-dateutil==2.4.0
pytz==2014.10
requests==2.5.1
twilio==3.6.15
git+git://github.com/dbinetti/django-noncense.git
boto==2.35.2
django-storages==1.1.8
waitress==0.8.9
django-autoslug==1.7.2
